AddComponent와 서버 연결
