+++ src/pages/Product/IndexPage.tsx
@@ -4,7 +4,7 @@ import { Navigate } from 'react-router-dom';
 function IndexPage() {
   
   return (
-      <Navigate to={'/product/list'}></Navigate>
+      <Navigate to={'/products/list'}></Navigate>
   );
 }
