+++ src/pages/Product/ProductAddPage.tsx
@@ -1,4 +1,4 @@
-import ProductAddComponent from "../../components/product/ProductAddComponent.tsx";
+import ProductAddComponent from '../../components/products/ProductAddComponent.tsx';
 
 
 
@@ -7,10 +7,7 @@ function ProductAddPage() {
 
   return (
     <div>
-      <div>
-        <div>Product Add Page</div>
-        <ProductAddComponent></ProductAddComponent>
-      </div>
+          <ProductAddComponent></ProductAddComponent>
     </div>
   );
 }
