+++ src/pages/Product/ProductListPage.tsx
@@ -1,4 +1,4 @@
-import ProductListComponent from '../../components/product/ProductListComponent.tsx';
+import ProductListComponent from '../../components/products/ProductListComponent.tsx';
 
 function ProductListPage() {
